Refactor: admin Filter on community Board & Tag CD (#219)
